assigning graph name to node name on update re: #559
should we allow the name of the root node of a graph to ever be different from the name of the graph itself?

Currently these are distinct values ( and ), and are assigned independently in the UI… I’m wondering if perhaps we should just always keep these two values in sync (and, thus, [potentially] remove  entirely….).